fix create comment bug

+++ src/controllers/likeController.js
@@ -2,8 +2,12 @@ import likeService from "../services/likeService.js";
 
 async function createOnEntry(req, res, next) {
   try {
+    const currentUser = {
+      id: req.user.id,
+      role: req.user.role,
+    };
     const { userId, entryId } = req.params;
-    const newLike = await likeService.createOnEntry({ entryId, userId });
+    const newLike = await likeService.createOnEntry({ entryId, userId, currentUser });
     res.status(201).json(newLike);
   } catch (error) {
     next(error);
@@ -11,8 +15,12 @@ async function createOnEntry(req, res, next) {
 }
 async function createOnComment(req, res, next) {
   try {
+    const currentUser = {
+      id: req.user.id,
+      role: req.user.role,
+    };
     const { userId, commentId } = req.params;
-    const newLike = await likeService.createOnComment({ commentId, userId });
+    const newLike = await likeService.createOnComment({ commentId, userId, currentUser });
     res.status(201).json(newLike);
   } catch (error) {
     next(error);
@@ -48,8 +56,12 @@ async function getById(req, res, next) {
 
 async function deleteById(req, res, next) {
   try {
+    const currentUser = {
+      id: req.user.id,
+      role: req.user.role,
+    };
     const { likeId } = req.params;
-    await likeService.deleteById(likeId);
+    await likeService.deleteById({ id: likeId, currentUser });
     res.status(204);
   } catch (error) {
     next(error);
